+++ archetypes/api/_APPNAME_/_APPNAME_-api/src/main/java/applica/_APPNAME_/api/Application.java
@@ -17,6 +17,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration;
+import org.springframework.boot.autoconfigure.mongo.embedded.EmbeddedMongoAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -38,7 +39,11 @@ import java.util.Date;
 @ComponentScan("applica._APPNAME_.data.hibernate")
 @ComponentScan("applica._APPNAME_.services")
 @ComponentScan("applica._APPNAME_.api")
-@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class, MongoAutoConfiguration.class})
+@EnableAutoConfiguration(exclude={
+        DataSourceAutoConfiguration.class,
+        MongoAutoConfiguration.class,
+        EmbeddedMongoAutoConfiguration.class
+})
 public class Application {
 
     static {