+++ archetypes/api/_APPNAME_/_APPNAME_-api/src/main/java/applica/_APPNAME_/api/configuration/ApplicationConfiguration.java
@@ -18,27 +18,26 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 /**
  * Created by bimbobruno on 14/11/2016.
  */
 @Configuration
-@Order(1)
 public class ApplicationConfiguration extends WebMvcConfigurerAdapter {
 
     private Log logger = LogFactory.getLog(getClass());
 
-    @Autowired
     private OptionsManager options;
 
     // FRAMEWORK GENERAL BEANS
 
     @Bean
     public OptionsManager optionsManager() {
-        return new PropertiesOptionManager();
+        options = new PropertiesOptionManager();
+        return options;
     }
 
     @Bean
@@ -80,15 +79,13 @@ public class ApplicationConfiguration extends WebMvcConfigurerAdapter {
 
     /* cors configuration */
 
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        String origins = options.get("applica.security.cors.allowedOrigins");
-        for (String origin : origins.split(",")) {
-            registry.addMapping(origin);
-
-            logger.info(String.format("Added corse origin: %s", origin));
-        }
-
-        super.addCorsMappings(registry);
+    @Bean
+    public WebMvcConfigurer corsConfigurer() {
+        return new WebMvcConfigurerAdapter() {
+            @Override
+            public void addCorsMappings(CorsRegistry registry) {
+                registry.addMapping("/**");
+            }
+        };
     }
 }