+++ archetypes/api/_APPNAME_/_APPNAME_-api/src/main/java/applica/_APPNAME_/api/configuration/HibernateConfiguration.java
@@ -18,7 +18,6 @@ import org.springframework.core.annotation.Order;
 
 
 //@Configuration
-//@Order(4)
 public class HibernateConfiguration {
 /*
     @Bean