+++ archetypes/api/_APPNAME_/_APPNAME_-api/src/main/java/applica/_APPNAME_/api/configuration/MongoConfiguration.java
@@ -3,10 +3,7 @@ package applica._APPNAME_.api.configuration;
 import applica._APPNAME_.data.mongodb.UserDetailsMongoRepository;
 import applica.framework.CrudStrategy;
 import applica.framework.DefaultRepository;
-import applica.framework.data.mongodb.DefaultMongoRepository;
-import applica.framework.data.mongodb.MongoCrudStrategy;
-import applica.framework.data.mongodb.MongoHelper;
-import applica.framework.data.mongodb.MongoMapper;
+import applica.framework.data.mongodb.*;
 import applica.framework.data.mongodb.constraints.ConstraintsChecker;
 import applica.framework.data.mongodb.constraints.SimpleConstraintsChecker;
 import applica.framework.security.UserDetailsRepository;
@@ -18,7 +15,6 @@ import org.springframework.core.annotation.Order;
  * Created by bimbobruno on 14/11/2016.
  */
 @Configuration
-@Order(4)
 public class MongoConfiguration {
 
     @Bean
@@ -46,4 +42,6 @@ public class MongoConfiguration {
         return new DefaultMongoRepository();
     }
 
+    @Bean
+    public MongoEmbedded mongoEmbedded() { return new MongoEmbedded(); }
 }