+++ archetypes/client/app/js/config.js
@@ -1,6 +1,8 @@
-var serviceBase = "http://192.168.0.45:3000/";
+var serviceBase = "http://localhost:8080/";
 
 module.exports = {
     "service.url": `${serviceBase}`,
-    "login.url": `${serviceBase}auth/login`
+    "login.url": `${serviceBase}auth/login`,
+    "account.register.url": `${serviceBase}account/register`,
+    "account.recover.url": `${serviceBase}account/recover`
 };