+++ archetypes/client/run-ios.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+aj build --platforms ios
+aj watch --platforms ios &
+
+wait