+++ archetypes/client/run-web.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+aj build --platforms web
+aj watch --platforms web &
+
+cd platforms/web
+bower install
+npm install
+gulp run &
+
+wait